initial commit for planner
I will change the design of the prioritized_tasks list, since Meta Bind Lists are problematic in the sense that I cannot force the UI to, say, only allow (or disallow) moving items up and down for tasks and chores.

This will be added soon.  It will be replaced with a Meta Bind Select and "move up"/"move down"/"edit" Meta Bind Buttons with JS actions.  For this to work I will need to use JS Engine's reactive script component to update the rendering for the Select.

Additionally, I will need to add a Meta Bind textArea (for writing a task down) and a Meta Bind Button with a JS action to add that task to the normal tasks list.  Ditto for chores.